Switch from s.resources to s.resource_bundle for MapboxCoreNavigation.podspec
This means that the privacy manifest (.xcprivacy) file is not copied to the main bundle by apps using the SDK with Cocoapods, which was causing build failures due to interfering with the root project's .xcprivacy file.

For statically linking, this commit also manually includes the MBXinfo.plist from MapboxNavigation and MapboxCoreNavigation, since the SDK throws an error otherwise.